YQL-19747: Split files
